requirements.txt
torch
torchvision
torchaudio
tqdm
numpy
kornia
more-itertools
Pillow
cupy-cuda101
random